FW: move auto_fd class to sandstone_utils.h
Right next to AutoClosingFile file.
